ml-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
transformers==4.36.0
torch==2.1.1
nltk==3.8.1
spacy==3.7.2
redis==5.0.1
pymongo==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
httpx==0.25.2
